requirements_e2e.txt
splinter==0.6.0
# nose-randomize
nose==1.3.1
randomize==0.9
PyHamcrest==1.8.0
